// src/app/shared/interfaces/subjectivities.ts
export interface Subjectivities {
    isIncluded: boolean | null;
    subjectivityCode: number | null;
    sortSequence: number | null;
    ysnDefault: number | null;
    ysnAutoSelect: number | null;
    ysnDeletable: number | null;
    subjectivityDesc: string | null;
    description: string | null;
    sequence: number | null;
    isUserDefined: boolean | null;
}